added "rerunable" to types of jobs allowed in queue disallowed_types acl
pbs_server now keeps a list of array_job_list structs, one for each job array
this struct contains a tlist_head for a list of all jobs that belong to that particular array.  Each job also has a pointer to this struct, and a list_link so it is linked into the list.