added array.h, req_deletearray.c and ran autoreconf
added stub code for array deletion request.  an array id can be passed to qdel and the request gets dispached to the req_deltearray() function. Currently the request is rejected with amessage that states that whole array deletion is not yet implemented

this stub will be filled out soon to delete an entire array with one request