Quiet thread lock order violations for tasks
The proper lock order for tasks is to acquire the alltasks mutex and
then the mutex for the specific task.
When a task is created, its mutex is locked. Then, trying to add
it to the proper task list acquires the alltasks mutex, which is an
ordering violation. This can't cause a deadlock, because nobody else
would be waiting on that task's mutex. Move the acquisition of the task
mutex to the functions that actually do the insertion. Now, these
functions expect the task unlocked and return it locked.
Also, there are a couple "trylock" situations that violate lock order,
but fall back to correct behavior if they cannot acquire the lock.
Ignore the trylocks when using Helgrind.