backported commit a50908295b8640dced3716257674d6d8c490c5e3 from the 4.1-dev branch.
This fixes a problem in the code to kill stray jobs. There are
cases where jobs are on the server but not the MOM. The MOM
returns an error message with the unknown job id but the server
could not match the message with the job id itself. Added logic
to parse the error message to find the job id so a comparison
could be made and job purged. TRQ-1624