Only send needed data over the wire for qstat (part 2)
Added a new batch request SelStatAttr. This is the same as SelStat
but it includes an attrl to limit what attributes need to be
returned.  qstat is currently the only client that uses this.  If
a new client is talking to an old server, it will fall back to
using SelStat.