This commit brings the trqauthd daemon from the 4.2.5 code which makes it possible to down trqauthd with a -d option. It also has changed the authorization protocol for clients by passing the host address in the protocol to the server.
The code for pbs_server has been changed to now look for the
host address of connecting clients along with the port when
authorizing connections. By matching address and port this
prevents a problem where client commands fail with an invalid
credential error because client commands from different hosts
use the same port value. pbs_server only matches the first port
in the server table and authorizes that connection. Any connections
after the first match are not authorized with just a port value.