TRQ-2620. Add more protection against pbs_server not being responsive.
When reading a request from the wire protocol, perform the first read with a short timeout
(5 seconds). If this times out only try to read again if the threadpool isn't considered
too busy. This prevents pbs_server from locking up entirely when it can't read requests
from the wire protocol.