TRQ-2434  Stray jobs on nodes. second fix
Added code to the scan_for_exiting portion of the code
to wait for all sister nodes to report in that they
had received a job termination from the Mother Superior.
This alleviates the problem of sister nodes not clearing
before the Mother Superior sends its obit to the server.