TRQ-2660 - TORQUE not cleanup up jobs correctly after a qrerun
When a mom sends its obit when a job is complete it still
needs some time to finish cleaning up the job. If a qrun
is executed immediately after a qrerun the job is still
on the MOM. We were not rejecting a qrun request in this
circumstance. We now do.