TRQ-2664. Fix a scenario will a job will be freed twice.
A close connection function was registered for queuing a job on the mom. This made possible a
scenario where if reply_jobid() failed at the end of a queue job request we would free twice -
first in the close connection function and then immediately following with a call to
mom_purge_job().