TRQ-2670 Server keeping jobs on node after job has been requeued
This fixes a preemption issue where a job is running and then
a call is made as follows:
   qrerun <jobid> && qrun <jobid>

The problem is that the job succeeds in running but it shows as
running on more nodes than it should. It also leaves the job
in a EXITING state on the MOM from which the job was preempted.
This is now fixed.