TRQ-2363. Add the ability to requeue all running jobs without contacting the moms.
qrerun all - which previously returned an error - now requeues all running jobs without
contacting the moms. Because this is a very serious and not-really-undoable operation
it can only be done by a manager and must receive confirmation in order to be performed.
The idea is to allow a manager to requeue all running jobs if for some reason the entire
cluster has become incapacitated.