TRQ-2632 pbsdsh requires FQDN even if other elements don't
fixed a problem with pbsdsh where it was given short names
for hosts in the PBS_NODEFILE but all hosts would resolve
to their fully qualified name with the tm APIs. This made
it so hosts would not match using the pbsdsh -h option.
We now look for the short name in the FQDN and make it
a match