TRQ-1260: qdel clientretry support
Added clientretry feature as it works in qsub:
* -b <num> command line argument support;
* CLIENTRETRY configuration option support;
* PBS_CLIENTRETRY environment variable support.
Clientretry allows try to connect to the server in a loop during the
given time period.

Added 3-times retry to send the general command (delete task) to the
server as it implemented in qsub, see TRQ-2417.

Added unit test.

Removed server code duplicate from qsub.