TRQ-2761: prevent resetting last status sent timestamp to 0
If torque MOM needs to force status update message sending in some cases
(if if pass not more than 1/3 of configured interval) it moves the last
status sent timestamp so the next status update will be sent after 1/3
of the interval. In other cases (if it passed more than 1/3 of the
interval) mom resets the last status update timestamp to 0 that makes
impossible to determine the actual last message sent time.

Updated the forcing function algorithm to never reset the timestamp to
0. Now it just moves the last timestamp like it's time to send the
update (time_now - update_interval).