Github issue #289. Correct node mic counts.
User mangalapandi reported this issue and suggested this fix, I'm adding a unit test
as well.
Essentially, the counts for the number of mics which are reserved for use on a node
are never being increased and the number of usable mics on a node are never being
increased (both are being decremented) so the counts aren't correct. This is now
resolved.