TRQ-2999 Job going into from 'C' to 'E' state after pbs_server restart and then disappears after second restart
The problem with this bug is that the job gets set to
a state of completed and a substate of aborted. When
the job is recovered TORQUE looks at the substate and
not the state and then puts the job into the exiting
state. The fix takes into account the state and changes
the substate to complete as well on restart.