TRQ-3049 Rolling upgrade causes some jobs to lose exec_port and total_runtime
This is not an upgrade problem. The problem was because the
total_runtime attribute was not getting written to the .JB
file so when pbs_server was restarted there was no total_runtime
attribute to copy into the job. Code was added to make sure
the ATR_TYPE_TV would be written on a jobs .JB file.