TRQ-2947. Fix a race condition on deleting jobs which are failing to start.
When a job fails to start, under certain conditions it will be requeued. If you delete a job during
the window of time where it fails to start, it will show up as requeued instead of completed.
This is now fixed by setting a flag on the job to say it is being deleted and then ignoring the
requeue exit code when it is received.