TRQ-3023. Retry failed commits no matter what error ALPS returns.
We are able to consistently reproduce a failure when confirming ALPS reservations if the
confirmation is done too quickly after the reservation is made. ALPS was also returning a
permanent failure because the reservation "didn't exist." The code now retries the
confirmation no matter what and it avoids the issue.

Conflicts:

	CHANGELOG