TRQ-3087 TORQUE server deleting jobs after pbs_server restart
When a job goes to a completed state we decrement the number
of queued jobs. The problem was that when we restarted pbs_server
we did not take that into account when recovering the completed
job and incremented the queue. This fix checks for completed
jobs at start time and does not increment the number of queued
jobs.