TRQ-3141  NOAA - $prologalarm not working
fixed a problem that was introduced in 5.0 during a refactor
of check_exiting_jobs. A timeout of 30 seconds was applied
the exiting job which caused it to end prematurely. I added
back the check for the default timeout as set in TJobStartTimeout
plus a check for $prolagalarm which sets pe_alarm_time and I use
the lesser of these two values for the timeout.