TRQ-3142  Increase maximum time for $prologalarm
In examine_all_running_jobs there was a loop that would
allow a job to run through it 10 times before it would then
kill a job because the assumption was that the job was hung.
This loop took about 540 seconds to go through. It causes a
problem with long $prologalarm times when running parallel
jobs because it causes the sister node to kill the job before
the mother superior is finished running the prologue. We now
wait until the default of 300 seconds before exiting or the
value of prologalarm which ever is larger.

Even with the expanded time the value set for prologlarm needs
to be longer than what is expected for a prologue script to run.
I found that when the prologalarm time on all the MOM nodes was
the same as the prologue script the sister not still killed the job.

Conflicts:
	configure.ac