Fix an issue with reporting the core count incorrectly.
This was due to os numbering not working as expected. On my laptop -
intel i5 - the first core has the os index of 0 and has the cpuset of 0
and 1. The second core has the os index of 1 and has the cpuset of 2 and
3. Torque previously expected the first cpuset index to match the os
index of the core. This changeset removes that expectation.