TRQ-3220 Torque auto_node_np does not override internal proc count of node
There is a global variable srv_clnodes which is a count of all of the
execution slots in the cluster. This value was not getting updated so
on a single node it is easy to expose the problem. Torque does a check
to see if it is even possible for the job request to every run. If the
ppn requested is larger than srv_clnodes then it will fail. Because
this value wasn't updated then the request was too large.