TRQ-3366 McGill pbs_server slow to respond to qstat commands
Fixed a deadlock when calling job_id_exists.

Also fixed a network live semi-lock in on_job_rerun.
For two cases we were not letting the job mutex lock
free when calling out the the MOM. When the MOMs were
unresponsive this caused a quasi deadlock situation
on pbs_server.