TRQ-3417 Torque now always fails to create cgroups with multi-req jobs
When qrun or Moab send in a node spec to Torque for the
-L qsub option it will have several hosts separated by
a + or a |. The + indicates the same req on separate hosts.
The | indicates the start of a new req. In the -l request the
code was written so that every entry separated by a + or |
were also a req. It is easy to have more old style reqs than
new style and so the job fails because it tries to access
new reqs that don't exist. This change maps the old style to new style