TRQ-3456 Empty Task Cpuset Files (due to possible race condition) (NUMA)
This is probably not a race condition as described in the ticket.
More than one cgroup can have the same cores. cpusets only
restrict where processes in a cgroup can run. It does not
restrict other cgroups from using the same cores. There are flags
that will allow this but we are not currently using them.

The function that set the cpuset returns a success or failure
code but it was not read. This check in reads the return code
and reports it back up the stack. On failure the user will know
the job did not start and it can be investigated.