TRQ-3517 pbs_server -d PBS HOME Directory not working correctly.
The PBS_ENVIRON variable was the only variable used for the pbs_environment
file. PBS_ENVIRON is set at configuration time so the -d option
cannot be used if the $TORQUE_HOME directory is moved.

We now mark if the -d option is used and if so we set the pbs_environment
variable in the code instead of using PBS_ENVIRON.