TRQ-3511
Remove the gperf build from the Makefile.am for drmaa.
Used the generated hash source code and modified it to
not use the inline code. The reason for this is because
pbs_attrib_lookup and drmaa_attrib_lookup would not get
included in the libdrmaa when inlined. Also did and extern "C"
for these functions to prevent c++ name mangling.

Conflicts:
	src/drmaa/src/drmaa_attrib.c
	src/drmaa/src/pbs_attrib.c
	src/drmaa/src/submit.c
	src/drmaa/src/wait.c