TRQ-3575. Make the mom cleanup the job without pbs_server's directions.
The mom now has all of the information it needs to clean up a job, and
it does so once the job ends. After successfully sending the obit, the
job is deleted.

pbs_server also distinguishes between older moms that need the direction
and newer moms that don't, and acts appropriately.