TRQ-3631 All processes are put into the first task cgroup on the sister nodes under openmpi
Added code for new tm api tm_spawn_user_daemon.

This API will spawn a user daemon process and add it to the root cgroup
of a job. This is in place to take care of a problem where OpenMPI
calls tm_spawn to launch its daemon and it is put into the first task cgroup
of the job. All subsequent processes end up there as well.