TRQ-3619. Make Torque able to handle very large job array submissions.
This is done by instantiating a limited amount of jobs at a time. Adding
the -i switch to qsub specifies a limit on the amount of array subjobs
that can be idle at a time. If -i is not specified, the admin may set a
limit, and Torque will default to 300 if no one specifies a limit.