TRQ-3752. Make Torque set the more lenient memory limit when multiple are specified for cgroups.
If a job is submitted with -l mem=X,pmem=Y,vmem=A,pvmem=B, the memory
limit will now be set to the higher of mem and pmem, and the swap limit
will be set to the higher of vmem and pvmem. In both cases we calculate
which will be the higher per-process limit.

Conflicts:
	src/lib/Libattr/complete_req.cpp