I have updated the configure logic to reject anything other than the PMIx v1.1.5 (or something further along the v1 series, should we release something) as that is the only release out there today. Future releases can be supported with the appropriate #if checks for PMIX_VERSION in the code.
I found a race condition in the PMIx integration that registers nspaces and clients. The PMIx APIs for that purpose are non-blocking, but they must complete prior to you fork/exec'ing any local clients. Otherwise, a connection request that arrives at the PMIx server prior to completing the registration functions will be rejected.

I don't know your threading model enough to figure out how to implement the wait function, so I left some comments (they start with "RHC" so you can search for them) that explain what needs to be done. Hopefully, those will be clear enough to get you thru the problem. If not, feel free to holler.