Let pbstop parser read consecutive index jobs
The pbstop contrib tool parses the output of 'qmgr -l -n' and
uses the information to display a grid of the cluster which shows
where jobs are running.

It seems that starting with commit 15737530 the format of the job
list returned by pbsnodes/qmgr changed so that multiple core jobs
was condensed and this borke the pbstop parsing algorithm.  This is
a simple patch that will recognize the new format.

The format for the "jobs" attribute of a node was:
    0/JOB1, 1/JOB1, 2/JOB1, 3/JOB2

Now it is:
    0-2/JOB1, 3/JOB2