Prevent job "Variable_List" from being lost between submission attempts
Previously, if a job submission through qsub did not succeed on first
attempt, then the ATTR_v attribute would be lost on the next
attempt to submit the job to the server.

This patch attempts to fix that by not deleting the attribute when
encoding the job attributes to send to the server.  When encoding
attropl's for the server, the ATTR_v is skipped since it seems that the
server expects the ATTR_v data after it has received both the job and
job resource data.